Move block production to gloas file (no logic change).
