+++ beacon_node/beacon_chain/src/block_production/mod.rs
@@ -1,0 +1,1 @@
+mod gloas;
