+++ beacon_node/beacon_chain/src/lib.rs
@@ -1,7 +1,6 @@
 pub mod attestation_rewards;
 pub mod attestation_simulator;
 pub mod attestation_verification;
-mod beacon_block;
 pub mod beacon_block_reward;
 mod beacon_block_streamer;
 mod beacon_chain;
@@ -10,6 +9,7 @@ pub mod beacon_proposer_cache;
 mod beacon_snapshot;
 pub mod bellatrix_readiness;
 pub mod blob_verification;
+mod block_production;
 pub mod block_reward;
 mod block_times_cache;
 mod block_verification;
@@ -24,7 +24,6 @@ mod early_attester_cache;
 mod errors;
 pub mod events;
 pub mod execution_payload;
-pub mod execution_payload_bid;
 pub mod fetch_blobs;
 pub mod fork_choice_signal;
 pub mod fork_revert;
