+++ beacon_node/beacon_chain/src/pending_payload_envelopes.rs
@@ -39,6 +39,7 @@ impl<E: EthSpec> PendingPayloadEnvelopes<E> {
 
     /// Insert a pending envelope into the cache.
     pub fn insert(&mut self, slot: Slot, envelope: ExecutionPayloadEnvelope<E>) {
+        // TODO(gloas): we may want to check for duplicates here, which shouldn't be allowed
         self.envelopes.insert(slot, envelope);
     }
